remove oudated functions
